qmail-queue patch to look for environment variable QMAILQUEUE
Bruce Guenter has written a patch which causes any program that would run
qmail-queue to look for an environment variable QMAILQUEUE. If it is present,
it is used in place of the string "bin/qmail-queue" when running qmail-queue.
This could be used, for example, to add a program into the
qmail-smtpd->qmail-queue pipeline that could do filtering, rewrite broken
headers, etc.
See http://untroubled.org/qmail/ for more information.

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>